requirements.txt
streamlit>=1.38.0
pandas>=2.1.0
openpyxl>=3.1.2
google-generativeai>=0.7.0
python-dotenv>=1.0.1
pytest>=7.4.0
rich>=13.7.0
sqlalchemy>=2.0.30
scikit-learn>=1.3.0
